Remove meta-clang from arago-master-* builds
Initial tests on master-wip are working.  We need to remove it so that
we can roll out the changes.

With the addition of clang in oe-core, the need for the meta-clang layer
is dwindling.  Take this time to remove it for testing so we can verify
if we can remove it for all master configs.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>